@@ -1,6 +1,6 @@
 package com.vector.appupdatedemo.http;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.lzy.okgo.OkGo;
 import com.lzy.okgo.model.Progress;
