@@ -17,7 +17,7 @@
 package com.vector.update_app;
 
 
-import android.support.v4.content.FileProvider;
+import androidx.core.content.FileProvider;
 
 public class UpdateFileProvider extends FileProvider {
 }
